Refactor & Fixups & Normalizations
- Renaming for more consistent the Scoped Wrappers
- Normalizing ETL String Creations
- Fixing Up Address Base Miscalculation into SetMemoryX section

+++ src/File.cpp
@@ -166,7 +166,7 @@ int FilePrint(File* stream, const char *format, ...) {
     va_list args;
     va_start(args, format);
 
-    size_t formatSize = vsnprintf(NULL, 0, format, args);
+    int formatSize = vsnprintf(NULL, 0, format, args);
 
     if(formatSize == 0)
         return 0;
